Update patch set 1
Patch Set 1:

> Jamo, the issue in your case is that you are using a distribution
 > created in jenkins sandbox and current script is not ready for
 > that. Your patch may fix something but may alco break the
 > distribution coming from autorelease so we have to be careful here.
 > Let me check at Vratko's run too and come with something safe.

it's not checking anything specific to sandbox, but it is a result
of me trying to test with a karaf4 distro that was created by an
autorelease job from the sandbox.

anyway, I put a CSIT job in the sandbox with this one line change
and run with the karaf3 distro [0] and karaf4 distro [1] created
from the autorelease sandbox job [2] and both seemed to deploy the
controller fine. obviously, I get the sense I'm missing something
here, but I am in a general state of confusion about everything 
right now.

[0] https://jenkins.opendaylight.org/sandbox/job/aaa-csit-1node-authn-only-carbon/2
[1] https://jenkins.opendaylight.org/sandbox/job/aaa-csit-1node-authn-only-carbon/3
[2] https://jenkins.opendaylight.org/sandbox/job/autorelease-release-notests-carbon/1/artifact/integration/distribution/

Patch-set: 1