Update patch set 2
Patch Set 2:

Hi Anil,

Is it possible to get the logs for yum -y install and apt-get install.
As the errors are related to some of the dependencies which we are installing through our compilation script, it would be of great help in debugging if we can get these logs and the list of the installed packages as well.

Thanks
Rahul

Patch-set: 2