Fix issue caused by process substitution on sh
The change fixes the issue 'syntax error near unexpected token `<''
which is caused because the while loop and find uses process
substitution feature which is not standard on all shells. The
default shell used to run a builder is 'sh'. To fix the issue
invoke the code from a script using bash.

Change-Id: Icac081a6ca1aa19cdcef8202cd943348e6dbc513
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>