Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Since we are doing a pip install git-review here: https://github.com/opendaylight/releng-builder/blob/master/jjb/opendaylight-infra-push-gerrit-patch.sh#L5
 > do we still need this in the java-builder.sh ?
 > 

Not sure if that's the only script that uses git review but it would be good to have git review installed as default on the system since it's useful. In fact we should probably have git review inside of basebuild.

If someone wants a newer git review than what the package manager provides they can virtualenv it.

Patch-set: 1
Label: Verified=0