Use -m pip from python instead of pip directly
When the workspace is too long, as can happen with long
job names, there is trouble running commands in the
virtualenv.

this is trying to find all the places this would be
needed, but specifically to address a few current
problems like [0] which sometimes completes but cannot
push logs because the job name is too long.

The idea came from shague [1] which came from
openstack [2].

[0] https://jenkins.opendaylight.org/releng/user/jluhrsen/my-views/view/netvirt%20csit/job/netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-snat-conntrack-carbon/
[1] https://git.opendaylight.org/gerrit/#/c/51699
[2] https://review.openstack.org/#/c/399754

Change-Id: I9e99893d8de69ef0eb7b36309983cedc0befb549
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>