Update patch set 2
Patch Set 2:

Is this the right place to put this? basically it will add this setting to every job and call to mvn.

If we only want to disable it for distribution then we should probably configure it in more specific jobs?

Patch-set: 2
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0