Change array in loop bashism to satisfy shellcheck
Warning: I'm not totally sure this is right, but it seems to be the
recommended way to keep the same behavior we had but be explicit that
we want it. If shellcheck was actually catching a bug and we don't want
the old behavior, we should do the loop recommendation.

https://github.com/koalaman/shellcheck/wiki/SC2199

Change-Id: Ibcb2e88ad4df7d1789747b137e2135154ca8435a
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>