Update patch set 2
Patch Set 2: Code-Review+2

Yes this is the correct logic. Thanks!

The difference between [@] and [*] as far as I can  tell is.

array("a" "b" "c")

with [@] it loops like:
a
b
c

with [*] it doesn't loop and gives you a single space separated text:

a b c

Patch-set: 2
Label: Code-Review=+2