Update patch set 1
Patch Set 1:

> Jamo, you were the one who added this previously. Do you recall
 > which jobs used this file if any?
 > 
 > I'd like to stop archiving in Jenkins if possible so if we can
 > identify which jobs are using this we can look into a solution
 > where they can pull from Nexus instead.

Thanh,

it was for this TODO [0] which is still pending. The idea was that 
the distcompare script which is run automatically could generate 
it's own accurate list of current projects instead of using the
current static list of projects.

this is not a priority for me currently, but hopefully I can get to
it at some point. It doesn't matter where the dependencies.log 
goes, just that it goes somewhere predictable.

JamO

Patch-set: 1