Update patch set 1
Patch Set 1:

Is there a link to the patch for int/test side of things?

One issue I have with this is it seems to be specific to int/test only and if other projects / jobs want to also push data the would need to somehow pull the script from int/test but I haven't seen what the python script looks like yet. I'd like to see that before I can comment further.

Patch-set: 1
Label: Verified=0