Update patch set 4
Patch Set 4: Code-Review+2

so we can merge this, as it seems right technically, but once we merge it we have to make sure the images are updates. Any jobs that
is going to use include-raw-build-deb.sh will no longer get the 
packages until the images are updated.

so +2 from me, and when you are able to quickly update the images you
can merge and go for it.

Patch-set: 4
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2