Update patch set 22
Patch Set 22:

> I'm disappointed my comment about validating that every RPM has an
 > associated SRPM was ignored. A cp of 2 directories does not
 > guarentee we have an SRPM for every RPM.

I specifically address it

> "We're not actually building from source, just repackaging the outputs of other builds. Building from source would be awesome, but afaik a *massive* TODO."

To say it again/differently, we're just wrapping the pre-built tarballs in RPMs, not actually doing a from-source build. With a real from-source build, you would get an SRPM with the sources and a noarch RPM with the binaries and we'd need to distribute both as you said. With our build, both RPMs just contain binary tarballs.

Patch-set: 22