Update patch set 22
Patch Set 22:

> I guess I just didn't fully understand how these things are built.
 > As long as we can point validly back to a source then I guess it's
 > fine. If all it's doing is constructing from our official release
 > tarballs then we can say the source can be pulled from the git
 > release tags of autorelease.
 > 
 > Thanks for bearing with me on this one.

Np, thanks for making sure we're doing the right thing :)

I'd love to eventually do from-source builds for the RPMs, but it's really complex for Java projects (do we package 1000s of 3rd party JARs from-source too? do we have some sort of exception to not build them from source, cache in a build system?). It's a super-long-term-wish.

Patch-set: 22