Repoint global-jjb to GitHub mirror
The Linux Foundation has started mirroring global-jjb to GitHub. As
such, we need to update the submodule checkout to use this resource
instead of the direct upstream Gerrit repository as it will allow LF to
perform systems maintenance on Gerrit without disrupting the downstream
projects that depend on the submodule.

Change-Id: I729356d40435f744fbee2eb5f34e364f7796131d
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>