Update patch set 5
Patch Set 5:

I guess I don't understand this script. If this is the script to data to ES server then it needs to be even more generic. For example releng does not have "testplans" but we would like to be able to push data to ES too. We cannot use this macro in that case. The macro should take an input (Filename) and push that data to the server.

If this is purely the script to generate the json data for integration then I guess this is fine?

Patch-set: 5
Label: Verified=0