Update patch set 3
Patch Set 3:

LF's goal is to stop allowing archives to Jenkins altogether as storage in Jenkins is limited whereas storage in Nexus his high. If you want to maintain the convenience one strategy is to figure out how to modify the Build description at the end of the build to include links to things in Nexus that you find convenient. We currently use it to link to the logs server but I think it shouldn't be too difficult to figure out how to make it append additional information somehow. Maybe Anil can look into it?

Patch-set: 3