Add 6Wind quagga to ubuntu16.04 and centos7 images
Description :
6wind/zrpcd is opensource to handle the the quagga framework from
SDN controller and it will install dependency packages based on
OS flavors like CentOS7 and Ubuntu version 14.04.*,16.04.* &17.04 along
with quagga packages and default version will be v1=L3VPN and v2=evpn

Issue with the dev scripts (as noted in thread) is now resolved on
zrpcd repository. The changes on zrpcd repo is baselined to tested
version (20170731).

Ref:
https://lists.opendaylight.org/pipermail/dev/2017-May/003724.html

Change-Id: I5f4ad020ff85c75141ad7ee87a551bb2d588040b
Signed-off-by: maheshl <mahesh.loni@altencalsoftlabs.com>
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>