Simplify autorelease to not use release tags
Release tags for carbon, nitrogen and future releases are set as
the stream and no longer uses the format `Carbon-SR1`. Only for boron
the same format is preserved if any hotfix has to be released.
Therefore, pass stream as the default release for all releases
except for boron. Also the scripts are modified to set the release tag
as $STREAM by default.

https://lists.opendaylight.org/pipermail/release/2017-July/011619.html
Jira: releng-323
Change-Id: I7ecc4d3962bab741bc31efa636beb986b4b15476
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>