Simplify autorelease to not use release tags
Release tags for carbon, nitrogen and future releases are set as
the stream and no longer uses the format `Carbon-SR1`. For boron
the same format is preserved. Therefore, the scripts are modified
to determine if the release tag passed from the job and use it,
else use stream as the release tag.

https://lists.opendaylight.org/pipermail/release/2017-July/011619.html
Jira: releng-323
Change-Id: I7ecc4d3962bab741bc31efa636beb986b4b15476
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>