Fix packaging archive publisher missing
The publisher archive-artifacts was removed in patch
Ib640a3528fb64fbf666df8f026f4e2313f290472. We're trying to discourage
the use of archiving artifacts to Jenkins. This is a temporary solution
to fix the builder-merge job but long term we need to move the archives
to logs.opendaylight.org if it's small log like data or Nexus if it's
large RPM packages that need to be saved.

Change-Id: I7c2e2597d1d2e77322d0ff5922954ab43ed7d71d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>