Update patch set 1
Patch Set 1:

So the reason it failed was the packaging patch was merged AFTER the patch to remove the macro. This caused a scenario where both patches passed verify but because we did not rebase it caused a post merge failure.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0