Hard-code openjdk8 for validate job
The JDK version does not really matter for validate-autorelease as the
correct JDK is configured by the project in their verify jobs. Typically
infra always wants to use the latest version anyway. This fixes the
warning in Jenkins that {jdk} is not found because
validate-autorelease's autogenerator job does not set a JDK.

Issue: RELENG-361
Change-Id: Idb71caf601d5f49a89afaf423573e78dff9ee72c
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>