Update patch set 3
Patch Set 3: Code-Review-1

(5 comments)

Rob, we are modifying the opflex jobs to release using a Nexus staging repo similar to how we release odlparent and autorelease projects. This helps streamline the release process and ensures your artifacts are pushed to a standard location (Nexus) rather than being archived on the log server.

Anil, Do we plan on updating the opflex jobs to do the full build in one job in later patches? Otherwise I think this job depends on the other jobs to run first to be successful right? or did I miss a step here?

Patch-set: 3
Label: Code-Review=-1