Update patch set 1
Patch Set 1:

> > did you miss one file?
 > >
 > > jjb/netvirt/netvirt-full-integration.yaml:      - shell:
 > > !include-raw-escape: include-raw-setup-docker.sh
 > > jjb/netvirt/netvirt-full-integration.yaml:      - shell:
 > > !include-raw-escape: include-raw-cleanup-docker.sh
 > 
 > Jamo, Daniel: These are under netvirt, actually this patch was
 > meant only for ovsdb. Please refer https://git.opendaylight.org/gerrit/#/c/61243/
 > which has required changes relevent to netvirt.
 > 
 > I like to keep these patches separate, since its a diff project.
 > Also, if it fails its easier to narrow down smaller changes and
 > revert it.

the problem is if we merge this one, but not 61243, netvirt will
start breaking. that's why I was thinking it's better to have them
all in this patch. OR, I suppose you could make the netvirt change
(if you need them to be separate) to be a child of this change.

Patch-set: 1