Update patch set 3
Patch Set 3: Code-Review-1

(5 comments)

Disabling shellcheck is lazy and defeats the purpose of having shellcheck in the first place. We should be resolving shellcheck issues unless there's absolutely a good reason to disable it.

The only ShellCheck rule I've seen that has to be ignored is SC1091 due to undetectable files that are created at runtime. In all other cases we should be striving to fix them. imo.

Patch-set: 3
Label: Code-Review=-1