Refactor OpFlex jobs
OpFlex project builds and archives dependencies (libuv, rapidjson
and ovs) on Jenkins.

Now the jobs are modified to build dependencies as a part of
the verify, merge and release jobs, since this takes additional
two minutes to build all three dependencies.

Remove jobs which build and archive dependencies that are no
longer required.

Move all the builders and scm macros into a opflex macros file.

Modify scripts such that dependencies are built based on the
version required for the stream.

Jira: releng-3
Change-Id: Icf6f8d1a16f7cebfb5fe37bf8c6f8f350d7eb9fa
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>