Update patch set 1
Patch Set 1:

(1 comment)

I had thought we decided we weren't going to do this.  The ODL build infra is already far slower than it needs to be for this and we're just making it worse.  I'll note that there are plans to add additional dependencies including VPP from fd.io that we'd also need to build, as well as eBPF toolchains.

There are already a lot of painful workaround we need to do to use the ODL infra, and of course you guys keep having to do weird exceptions because of OpFlex.

Perhaps this is a good time to just move OpFlex dev out of ODL entirely, since it seems to be causing problems for you guys (for reasons I don't understand; you keep making references to a "log server" but I don't understand why you archive artifacts to a log server).  I think this would allow you to focus purely on actual ODL and a completely Java-based env.

Patch-set: 1
Label: Code-Review=0