Create mininet image with OVS 2.6.1
Netvirt L2GW CSIT needs OVS 2.6.1 to get newer
version (1.6.0) of the hardware_vtep schema.

This newer version is needed so that we can set
the replication_mode of the logical switch to
"source_node", so that BUM pkts will be replicated
by the HWVTEP (or HWVTEP emulator, in CSIT tests).

I have tested HWVTEP csit with OVS 2.6.1 in sandbox.
Here is a link to those runs, build 28 is where
the test run starts with successful upgrade
to OVS 2.6.1.

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-hwvtep-1node-openstack-newton-upstream-transparent-nitrogen/

There are 8 failures, in the BGP VPN test suite,
but they are not due to this patch. These tests
fail because quagga bgp was not in the image.

Change-Id: Ie26617b8f168cf3afad63f06bd25da152ac4a342
Signed-off-by: Victor Pickard <vpickard@redhat.com>
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>
Signed-off-by: Victor Pickard <vpickard@redhat.com>