Update patch set 1
Patch Set 1: Code-Review+1

I think this is a good move. In my opinion if a project joins they should decide if their feature is compatible with all or only and we only test against one of them. It is easy enough for troubleshooting for them to make a temp job on the sandbox if they want to test only.

Patch-set: 1