Update patch set 1
Patch Set 1:

> Depends what you mean by "history" because we keep logs on the log
 > server for 6 months. Yes the Jenkins side of things will disappear
 > with this patch but the logs server side of things will stick
 > around.

yeah, we will keep the logs in the logserver for 6 months, but this
made me think that we should probably send a mail on dev/release/integration about this when we are ready to merge, and then wait 
a week before merging. Just to give people the chance to change our
minds if they have a good reason.

Patch-set: 1