Refactor OpFlex jobs to push to nexus
OpFlex project builds and archives dependencies (libuv, rapidjson
and ovs) on Jenkins.

Now the jobs are modified to build and push the dependencies to
nexus thridparty repository. The verify, merge and release jobs are
modified to download these dependencies from nexus while building
opflex artifacts.

The jobs are designed to take the version number of the dependency
to build.

All the scripts are moved into a common `scripts`.
All builders and scms are moved to opflex macros file.

Jira: releng-3
Change-Id: Icf6f8d1a16f7cebfb5fe37bf8c6f8f350d7eb9fa
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>