Update patch set 3
Patch Set 3:

As mentioned in my comment in RELENG-371. It seems this patch has taken a different path then what was suggested in the description. So I will ask it again here:

Do we want to fail the build if the image is not public or automatically set visibility?

There is a key difference here because if we automatically set visibility this means anyone can automatically add a new image to our supported image list without it being curated. Whereas the other way around prevents people from using uncurated images. I think clarifying this matters because in the first case we are saying that we will support any image that anyone decides to start using. Whereas in case 2 we are saying we will only support images that has went through a verification process and someone made the decision to support a particular version of the image.

Patch-set: 3
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0