Update patch set 2
Patch Set 2:

> Hi Jamo/Anil,
 > We need your input on this :
 > 1.thrift-version = 1 : Boron Release ( default)
 > 2.thrift-version = 2 : Carbon or Nitrogen ( without IPv6)
 > 3.thrift-version = 4 : Carbon or Nitrogen ( with IPv6)
 > 
 > Currently thrift version=1 is available in community with centos7
 > and Ubuntu16.04 image
 > 
 > thrift version=2 and 4 is still pending
 > is it possible to define image name based on streams ( boron or
 > carbon or nitrogen ) in yaml files or need to create a separate
 > yaml for each release. please let us know best approach to proceed
 > further on this.

we do something (well, *did* something) similar in netvirt. look
at these lines:

https://github.com/opendaylight/releng-builder/blob/master/jjb/netvirt/netvirt-csit-multi-openstack.yaml#L81-L98

before we had csit images based on mitaka and newton and depending
on the openstack branch we would pick the mitaka or newton image.
We've since moved on and no longer have mitaka in our csit, and
all the branches work with newton. But you can see how we were
controlling it.

that might be a good option here.

Patch-set: 2