Don't try/fail to push pkgs sandbox->Nexus
The sandbox Jenkins silo doesn't have permission to push to Nexus, so
jobs that would push packages try to do so and fail, marking the entire
job as a failure. This checks if in sandbox and doesn't try if so,
preventing false negatives.

Ideally we'd also host the packages on sandbox Jenkins, but I haven't
figured out how to do so.

Change-Id: If1d660e8462c2e459a8d2aa69c6ce99af387b293
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>