Update patch set 1
Patch Set 1: Code-Review-1

Jamo, take a look at this as a start to get rid of the no .stestr log. only need to look at the part where the scp to the node is node. I put the file in the tempest dir and it is found, but then testr looks for the path in the file  - and no matter what vale I put in there the tempest blows up.

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1