Fix L2GW CSIT
When SFC was added to NetVirt CSIT, L2GW CSIT stopped
working.

SFC installs python-neutronclient, and ends up getting
6.5.0. This somehow breaks L2GW CSIT, with error
that l2-gateway-xxx commands are not recognized.

Installing python-neutronclient 6.1.0 after installing
networking-l2gw and networking-sfc will get the L2GW
CSIT passing again.

But, L2GW CSIT should work with python-neutronclient
6.5.0, it does on my local setup. So, it would
appear some other python dependency is in play
here, just not sure what it is yet. Jamo, any
ideas?

Follow-up from discussions with Jamo (thanks!):
Robot runs in separate VM, and not on one of the
devstack nodes. So, the Openstack/Neutron CLI
python modules are loaded separately on the
robot vm, and CLI commands are executed on the
robot vm, instead of sshing to the devstack node.

This will work for now, but one alternative would
be to do a git clone of the source for the branch
under test, and install from there, instead
of hardcoding install.

Change-Id: I5d2c7cf522639bda4d1a68d5717f4b89aa15cff9
Signed-off-by: Victor Pickard <vpickard@redhat.com>
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>
Signed-off-by: Victor Pickard <vpickard@redhat.com>