Update patch set 3
Patch Set 3:

> Jamo,
 > Here is a hwvtep run with your patch set 3. This doesn't work.
 > 
 > What I don't understand, is why we even have to do an explicit "pip
 > install networking-l2gw", or even "pip install networking-sfc" in
 > the robot setup. I thought devstack would take care of this if the
 > local.conf is configured correctly, at least it does in my local
 > pike/carbon setup. Do you know why we have to do these pip installs
 > in the robot setup?
 > 
 > FYI, I started build 15 in sandbox without these pip installs,
 > which will probably fail, but I will debug from there.
 > 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-hwvtep-1node-openstack-pike-upstream-stateful-carbon/14/

for posterity, this patch wasn't used in your job #14 or #15. I think
you were trying to set the gerrit_refspec value with this patch, but
TBH I don't even know what gerrit_refspec refers to for CSIT jobs. We
have to push an updated job to the sandbox using this patch. you
can do that by having these changes locally and running the 
jenkins-jobs update command, or there is now a new method by just typing
a gerrit comment to update/create the job automatically from a releng/builder
gerrit review. see my earlier comment on ps3 for an example.

as for why we are installing these things explicitly, I do not know.
If devstack will do that for us, we can move that work there. Is that
just a setting in the local.conf? If so, we can create a patch to 
handle that. This "install-robotframework" script was probably just
a logical place to put packages that we wanted to have on the test
execution VM.

I'm also guessing that doing it in devstack would mean that it's doing
a git clone of whatever project, and checking out whatever $branch and
doing a pip install inside that repo. This install-robotframework
script is doing it from the python upstream releases. not sure if it's
different or not.

Patch-set: 3
Label: Verified=0