Update patch set 1
Patch Set 1:

(2 comments)

> (2 comments)

> I don't get why rtd jobs need a different patch but modifying python jobs can stay in this patch? If we're gonna make different patches based on topics then we need 3 patches for changes to rtd, python, and node.

Hey Thanh: I prefer keeping it as a separate patch since this change switches 3 jobs types to global-jjb including also removes the jobs from releng templates, and in case if there was an issue with one of the job type we may have to revert a jobs temporarily, it's easier to revert the change specific to the job in question. Hope this clarifies.

Patch-set: 1