Update patch set 2
Patch Set 2:

> Hey Thanh: I prefer keeping it as a separate patch since this
 > change switches 3 jobs types to global-jjb including also removes
 > the jobs from releng templates, and in case if there was an issue
 > with one of the job type we may have to revert a jobs temporarily,
 > it's easier to revert the change specific to the job in question.
 > Hope this clarifies.

Hey Anil, What I'm confused about is you're treating RTD jobs as a different job yet treating nodejs and python as the same job. This is the differentiation I don't understand. Why?

I'm fine with splitting out the job but we should be consistent with the statement here. If you want to make patches for different job types than NodeJS is not related to Python either. If this patch is about migrating the node job then I would expect a separate patch about migrating python jobs to be consistent with your statement that RTD jobs are different.

Patch-set: 2
Label: Verified=0