Add infrautils-ready feature to distro features
infrautils-ready will give us log messages indicating when
all features have been successfully installed and the overall
system is ready for further testing.

this patch simply adds that as a default feature in all of
the places we are bringing up the controller. What we do
with it after that is still work to be done.

Also, this feature is not there in carbon or earlier, so those
jobs cannot take advantage of it. This patch is not smart
in that it will still list it as a feature for a carbon
distro. It should not harm anything to do that.

Change-Id: I8f3b9b2d40edd482a45f091b30be7345f8b4e088
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>