Update patch set 1
Patch Set 1:

Yes, we can tweak the job as we go. I just want to get the job out there soon so that we can test the job itself.

My current thinking about the voting is we do want it to vote and override existing votes. If my thinking is correct for version bump patches distro-check and validate-autorelease are expected to fail anyway so we want to ignore them and this verify job overrides the regular verify job.

I'm hoping with this job we can have all the patches have a +1 vote by Jenkins for a version bump before we start merging the patch series.

I do see a possibility of projects using this job to bypass distro-check and validate-autorelease jobs unintentionally because those jobs don't run but considering this job is a very special case job maybe it should not be used that often.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0