Update patch set 1
Patch Set 1:

Just realized the OpenDaylight jobs unlike the global-jjb jobs checks for "recheck" anywhere in the comment so causes a verify to run. This will be fixed as we migrate over to global-jjb with more jobs. Although we should probably add "recheck$" to existing OpenDaylight jobs to ensure recheck only runs if the comment contains that at the end like we do with global-jjb.

Patch-set: 1