Update patch set 1
Patch Set 1:

Venkat, how much does this reduce the log sizes? Is this to help with the issue we hit when the output.xml is too big to be processed and log.html fails to be produced?

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0