Update patch set 1
Patch Set 1:

> Venkat, how much does this reduce the log sizes? Is this to help
 > with the issue we hit when the output.xml is too big to be
 > processed and log.html fails to be produced?

yeah, my guess is that this will remove all the logging we get in the
: FOR loops, which will reduce the size of output.xml overall. BUT,
I don't think we want that. I want to see the log output of each item
in the for loop.

I think if we bumped the size of our robot VM (cores and/or RAM) a
little bit, we'd be ok. I can easily generate the .html files on my
laptop, but it does take a bit of horsepower.

Patch-set: 1
Label: Verified=0