Update to use devstack tip of branch
The main change to latest devstack is that the computes need
to query the control node. The computes will poll for 60s and if
not found the compute stack fails. This patch does the following
to support the change:

1. Modify the devstack code to increase the 60s timeout to 30m.
In multiple tests the average stack time for a control node is 20m
and the computes 18m so 30m shoudl provide coverage.

2. Update the iptables handling. The previous run.sh iptables config
simply disabled the firewalld and iptables, but devstack restarts
iptables so the query from the computes to the control node would
fail. This change leaves iptables running but updates the rules
to open the OpenStack ports. This better matches a practical
deployment so it is good to have.

3. Add a new local.conf option: CELLSV2_SETUP=singleconductor.
Another change in devstack is that when the nova-compute is started
on the computes, it queries the control node for the cell1. The cell1
access in mysql is not set up until later in the control node bringup.
If the compute queries before that setup an exception is not caught
and the service exits - and therefore there is no hypervisor on that
node. The local.conf change removes that query.

4. Added some more conf file captures to the collect_logs() function.

Change-Id: I8a4c05cce3107f0525022dfe06c46a78cf3b78c9
Signed-off-by: Sam Hague <shague@redhat.com>