Bump stack timeout to 15 minutes
Also lower stack retries to 2. This ensures that the overall timeout of
the job does not increase too. This is an attempt to workaround RS
potentially taking too long to spin up servers as a reason the nodes are
not coming online.

Let us see if a 15 minute timeout improves the situation any.

Change-Id: I2ed403aab37233b2e69022daf7e0f58dff8fc3c9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>