Update patch set 4
Patch Set 4: Code-Review+2

> > > > Venkat, how much does this reduce the log sizes? Is this to
 > > help
 > > > > with the issue we hit when the output.xml is too big to be
 > > > > processed and log.html fails to be produced?
 > > >
 > > > yeah, my guess is that this will remove all the logging we get
 > in
 > > > the
 > > > : FOR loops, which will reduce the size of output.xml overall.
 > > BUT,
 > > > I don't think we want that. I want to see the log output of
 > each
 > > > item
 > > > in the for loop.
 > > >
 > > > I think if we bumped the size of our robot VM (cores and/or
 > RAM)
 > > a
 > > > little bit, we'd be ok. I can easily generate the .html files
 > on
 > > my
 > > > laptop, but it does take a bit of horsepower.
 > >
 > >
 > >
 > > Jamo,
 > >
 > > When we use flattenkeywords, I think it is a bit more efficient
 > >
 > > Please find sandbox run here
 > > https://logs.opendaylight.org/sandbox/jenkins091/netvirt-gvrangan-csit-1node-openstack-pike-upstream-stateful-snat-conntrack-nitrogen/9/
 > >
 > > Added the url which suggested this approach:https://github.com/robotframework/robotframework/issues/1833
 > 
 > Also, It will report if an iteration in FOR loop failed, it will
 > only flatten the success. All failures will be available in
 > log.html

My main worry was that happy path logging output would not be there. Sometimes
I need to drill back up in a suite and find some mac address or some port id
when figuring out what a later test case was looking at. It seems that we'll
still get all the logging output that comes anyway, but we will lose a lot of
the robot overhead logging. Looks good to me. here is an example of a : FOR
loop going through each instance and doing an "openstack server show":

with flattenkeywords:
https://logs.opendaylight.org/sandbox/jenkins091/netvirt-gvrangan-csit-1node-openstack-pike-upstream-stateful-snat-conntrack-nitrogen/9/log.html.gz#s1-s1-s1-t8-k17-k1-k6

and without:
https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-ocata-upstream-stateful-oxygen/254/log.html.gz#s1-s1-s1-t8-k17-k1-k6

it's not a whole lot different from what is displayed, but doing this on the
entire suite changed our final log.html file size to aprox 30% smaller than
without it. (41M down to 29M)

good work.

Patch-set: 4
Label: Code-Review=+2