Update patch set 2
Patch Set 2:

> what's more cumbersome:
 > 
 > - not getting a log.html and having to locally download the
 > output.xml and then running robot.rebot on it to see the
 > log.html local?
 > 
 > - having to look at the pages of module output when we are
 > debugging?
 > 
 > 
 > I don't care, thb. One thing though, is that this loopremover
 > addition didn't solve all the cases where we don't get the
 > log.html generated. Two other things will help. Less "Log"
 > statements in robot and a bigger VM (more RAM)

I am inclined to make it easier for the people debugging so that they actually debug. That means going to the old way and list out the individual models. If log.html is not there then they can do the manual steps but that is a low frequency.

This VM is already a 4gb/4cpu vm right? Pretty bad if we can't process the file within those specs. What is we don't let the pybot dump the log.html and then run the rebot in a following step? Maybe that will behave better since the pybot run is done?

We can also look again at suppressing the DeprecationWarning's since I think that is what is filling up the logs. I could not get it to work though.

Other thought is breaking the pybot call into per suite calls. I actually like this so we would have smaller files. We just need to pass args into pybot specifying the file name rather than letting it set them to default output.xml and log.html.

And last, I am working on the patch to use a combo control+compute which will drop those logs by 1/3. That may not help in the 3node tests since I think those are the real issue.

Patch-set: 2
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0