Update patch set 2
Patch Set 2: Code-Review+2

> > what's more cumbersome:
 > >
 > > - not getting a log.html and having to locally download the
 > > output.xml and then running robot.rebot on it to see the
 > > log.html local?
 > >
 > > - having to look at the pages of module output when we are
 > > debugging?
 > >
 > >
 > > I don't care, thb. One thing though, is that this loopremover
 > > addition didn't solve all the cases where we don't get the
 > > log.html generated. Two other things will help. Less "Log"
 > > statements in robot and a bigger VM (more RAM)
 > 
 > I am inclined to make it easier for the people debugging so that
 > they actually debug. That means going to the old way and list out
 > the individual models. If log.html is not there then they can do
 > the manual steps but that is a low frequency.
 
I know this was happening pretty much on every 3node job, but we
can merge this if it's causing pain in debugging 1node logs.

I'll merge

 > This VM is already a 4gb/4cpu vm right? Pretty bad if we can't
 > process the file within those specs. What is we don't let the pybot
 > dump the log.html and then run the rebot in a following step? Maybe
 > that will behave better since the pybot run is done?
 
This is just like our ODL crash. The process of robot.rebot is consuming
too much memory. I think doing it manually, or letting pybot do it is
just the same underlying tool so we'd hit the issue either way. I think
the problem is the whole output.xml and it's processing is done in memory
all at once and just a scale problem with the tool. More memory is needed
for bigger logs.

 > We can also look again at suppressing the DeprecationWarning's
 > since I think that is what is filling up the logs. I could not get
 > it to work though.
 
yes this would help

 > Other thought is breaking the pybot call into per suite calls. I
 > actually like this so we would have smaller files. We just need to
 > pass args into pybot specifying the file name rather than letting
 > it set them to default output.xml and log.html.

this is fine too. figuring out how to finish the work so the single
job looks like single robot suite at the end is what I don't know
how to do. this will be a bit of work to get right.

 
 > And last, I am working on the patch to use a combo control+compute
 > which will drop those logs by 1/3. That may not help in the 3node
 > tests since I think those are the real issue.

cool beans.

Patch-set: 2
Label: Code-Review=+2