Add new WIP job for netvirt csit
The purpose of this job will allow for unstable/new tests so
they can be run on a normal cadence while things are being
resolved.

One paradigm is to add tests before features and allow the
CSIT job to run and fail while feature dev is being worked
on. Assuming proper tests, the features can be validated
automatically with this kind of job

Change-Id: I429c5fb8c7da11943f51ff1ffbd7b44c50f5b77e
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>