Update patch set 3
Patch Set 3:

(1 comment)

Venkat, good stuff. Really like this.

looks like the log.html files are duplicated. In the root and then in the robot-plugin. it would be good to have only one copy in the robot-plugin so we keep everything tidy and organized.

We can also look at naming the files with the log or output before the extension, like 01_l2.log.html and 01_l2.output.xml so that way related files are next to each other in the listing if that makes it more readable.

Patch-set: 3