Update patch set 3
Patch Set 3:

> a few things to understand/figure-out:
 > 
 > 1) I don't see a single log.html get created. Can we still get
 > this?
Do we need one or can that be figured in a later patch? The smaller log.htmls load faster. It is a pain if wanting a quick look at what failed all in one shot.
 > 
 > 2) the resulting output.xml in the robot-plugin folder is not as
 > big
 > as it is when we are seeing the normal pybot process choke and
 > get killed. That's the condition we need to see if this patch
 > resolves. I think it happens when that file approaches 54M
All the files aggregated should end up about the same so not sure why we need to know this. As long as each file is less than 54M then we are good right? And that is what this patch does.
 > 
 > 3) If test case names must be unique we need to find a process
 > to guarantee that (non human process) because it's pretty
 > easy to write a new test suite and one of the test cases is
 > "Create Networks"
 > 
 > 4) Is the jenkins robot plugin just parsing all of the individual
 > robot.html files to know the pass/fail count? that's cool, if
 > so.
Looks like it since it had all the tests.

Patch-set: 3
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0